fix middleware

+++ routes/web.php
@@ -33,17 +33,17 @@ Route::get('/blank', function () {
 })->name('blank');
 
 Route::middleware('auth')->group(function() {
-    Route::resource('basic', BasicController::class);
-    Route::resource('driver', DriverController::class);
-    Route::resource('vehicle', VehicleController::class);
+    Route::resource('basic', BasicController::class)->middleware(['role:' . Role::ROLE_OWNER]);
+    Route::resource('driver', DriverController::class)->middleware(['role:' . Role::ROLE_ADMIN]);
+    Route::resource('vehicle', VehicleController::class)->middleware(['role:' . Role::ROLE_ADMIN]);
     Route::get('loan/approved', 'LoanController@index')->name('loan.index');
-    Route::get('loan/pending', 'LoanController@pendingLoan')->name('loan.pendingLoan');
+    Route::get('loan/pending', 'LoanController@pendingLoan')->name('loan.pendingLoan')->middleware(['role:' . Role::ROLE_ADMIN]);
     Route::get('loan/rejected', 'LoanController@rejectedLoan')->name('loan.rejectedLoan');
-    Route::get('loan/create', 'LoanController@create')->name('loan.create');
-    Route::post('loan/store', 'LoanController@store')->name('loan.store');
-    Route::delete('loan/destroy/{id}', 'LoanController@destroy')->name('loan.destroy');
-    Route::get('loan/owner/pending', 'LoanController@indexOwner')->name('loan.indexOwner');
-    Route::get('loan/edit/{id}', 'LoanController@edit')->name('loan.edit');
-    Route::put('loan/update/{id}', 'LoanController@update')->name('loan.update');
+    Route::get('loan/create', 'LoanController@create')->name('loan.create')->middleware(['role:' . Role::ROLE_ADMIN]);
+    Route::post('loan/store', 'LoanController@store')->name('loan.store')->middleware(['role:' . Role::ROLE_ADMIN]);
+    Route::delete('loan/destroy/{id}', 'LoanController@destroy')->name('loan.destroy')->middleware(['role:' . Role::ROLE_ADMIN]);
+    Route::get('loan/owner/pending', 'LoanController@indexOwner')->name('loan.indexOwner')->middleware(['role:' . Role::ROLE_OWNER]);
+    Route::get('loan/edit/{id}', 'LoanController@edit')->name('loan.edit')->middleware(['role:' . Role::ROLE_OWNER]);
+    Route::put('loan/update/{id}', 'LoanController@update')->name('loan.update')->middleware(['role:' . Role::ROLE_OWNER]);
     Route::get('loan/export', 'LoanController@export')->name('loan.export');
 });
